Updated namepsbt code for PSBT refactoring / V2.

+++ src/rpc/net.cpp
@@ -1167,7 +1167,7 @@ static RPCMethod exportasmap()
 
             UniValue result(UniValue::VOBJ);
             result.pushKV("path", export_path.utf8string());
-            result.pushKV("bytes_written", (uint64_t)node::data::ip_asn.size());
+            result.pushKV("bytes_written", node::data::ip_asn.size());
             result.pushKV("file_hash", HexStr(hasher.GetSHA256()));
             return result;
 #endif
